Better logging and error reporting for PushStatus

+++ com.archimatetool.modelrepository/src/com/archimatetool/modelrepository/workflows/DeleteBranchWorkflow.java
@@ -14,12 +14,11 @@ import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.dialogs.ProgressMonitorDialog;
 import org.eclipse.jgit.api.errors.GitAPIException;
 import org.eclipse.jgit.transport.PushResult;
-import org.eclipse.jgit.transport.RemoteRefUpdate;
+import org.eclipse.jgit.transport.RemoteRefUpdate.Status;
 import org.eclipse.osgi.util.NLS;
 import org.eclipse.ui.IWorkbenchWindow;
 
 import com.archimatetool.editor.ui.components.IRunnable;
-import com.archimatetool.editor.utils.StringUtils;
 import com.archimatetool.modelrepository.authentication.UsernamePassword;
 import com.archimatetool.modelrepository.repository.ArchiRepository;
 import com.archimatetool.modelrepository.repository.BranchInfo;
@@ -108,9 +107,24 @@ public class DeleteBranchWorkflow extends AbstractRepositoryWorkflow {
                 logger.info("Deleting remote branch: " + branchInfo.getLocalBranchName()); //$NON-NLS-1$
                 PushResult pushResult = utils.deleteRemoteBranch(branchInfo.getLocalBranchName(), npw, new ProgressMonitorWrapper(monitor,
                                                                                         Messages.DeleteBranchWorkflow_0));
-                logResult(pushResult);
                 
-                // Then delete local and tracked branch
+                // Logging
+                for(String msg : GitUtils.getPushResultMessageList(pushResult)) {
+                    logger.info(msg);
+                }
+                
+                // Get any errors in Push Result and throw exception
+                Status status = GitUtils.getPrimaryPushResultStatus(pushResult);
+                
+                if(status != Status.OK && status != Status.UP_TO_DATE && status != Status.NON_EXISTING) {
+                    String errorMessage = GitUtils.getPushResultFullErrorMessage(pushResult);
+                    if(errorMessage == null) {
+                        errorMessage = "Unknown error"; //$NON-NLS-1$
+                    }
+                    throw new GitAPIException(errorMessage) {};
+                }
+
+                // If OK, delete local and tracked branch
                 logger.info("Deleting local branch: " + branchInfo.getShortName()); //$NON-NLS-1$
                 utils.deleteBranches(true, // force the delete even if the branch hasn't been merged
                                    branchInfo.getLocalBranchName(),
@@ -134,17 +148,6 @@ public class DeleteBranchWorkflow extends AbstractRepositoryWorkflow {
         }
     }
     
-    private void logResult(PushResult pushResult) {
-        String messages = GitUtils.getPushResultMessages(pushResult);
-        if(StringUtils.isSet(messages)) {
-            logger.info("PushResult message: " + messages); //$NON-NLS-1$
-        }
-        
-        for(RemoteRefUpdate refUpdate : pushResult.getRemoteUpdates()) {
-            logger.info("PushResult status for " + refUpdate.getRemoteName() + ": " +refUpdate.getStatus()); //$NON-NLS-1$ //$NON-NLS-2$
-        }
-    }
-
     @Override
     public boolean canRun() {
         return currentBranchInfo != null &&
